refactor(tenant): extract full tier system (542 lines) to proximadb-tenant + fix GraphOperationsPort
Move the entire tenant tier DEFINITION system from src/catalog/tenant_tier.rs to the
foundation crate proximadb-tenant:
- Tier enum + impl (tier_row-based data lookups)
- TierSpec, TierSoftCaps, TierObjectEconomyConfig, ObjectEconomyQuantizationCeiling
- FeatureFlags, TenantTierRecord + impl
- tier_row + tier_config + parse_and_validate (config loading)
- BAKED_TIER_CONFIG_BYTES (include_str! path fixed for new crate location)
- BudgetDecision stays in root (store-layer type)

Deps added to proximadb-tenant: anyhow, serde_json, async-trait, tokio, tracing.

Also fixes the GraphOperationsPort impl: inherent methods return
Result<_, VectorDBError>, port expects anyhow::Result — fixed with ? conversion.

Resolves ALL 3 catalog deps (Tier + FeatureFlags + TenantTierRecord) for the
observability extraction. Combined with TenantStorageUsage + graph-model +
GraphOperationsPort: 8 of 9 observability outbound deps now resolved. Only
PlanOutput remains.
