+++ crates/foundation/proximadb-tenant/Cargo.toml
@@ -16,6 +16,11 @@ path = "src/lib.rs"
 [dependencies]
 proximadb-proto.workspace = true
 serde.workspace = true
+anyhow.workspace = true
+serde_json.workspace = true
+async-trait.workspace = true
+tokio.workspace = true
+tracing.workspace = true
 
 [lints]
 workspace = true
